src/icn/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorlog==4.8.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
paho-mqtt==1.5.1
pymongo==3.11.3
requests==2.25.1
urllib3==1.26.4
Werkzeug==1.0.1
